New query logic using Media and MediaList queries
This deprecates MediaListCollection query. Initially mendo will query
for MediaListCollection, save it to local database and then use the
archive file name from MComix to query from this local database and then
update the Media. This should work fine but there are two problems:
1. Keeping this local database in sync is a huge hassle, especially for
   a CLI program.
2. In order to search this local database mendo will likely have to
   implement a fuzzy search function somehow, this is bloated since the
   anilist API already provides a search function.
3. MediaListCollection will only get data from watching, reading,
   planning and so on, it will not retrieve an arbitary title user has
   never add to their list.
So in this commit mendo will just use anilist API search, this has one
downside of having to implement an extra MediaList query to get user's
progress, status and stuff to update (3 queries in total to update).
However, we can save the result of Media response to a file and look it
up next time to save time. This will save time on waiting for one
response from anilist.