Add get_media_id function
This function will first check if there is media_id of manga name from
local media data file (same dir as log file). If it exists then just use
that id instead. If it does not exist then it will query for that ID and
then append to the file. There is one problem that if the user messes
with this file then they will get the wrong result. May need to handle
this error.